Add IntervalClock component
Separate the clock from the intervalWorkout component
