+++ src/components/GeneralComponents/StartButton.js
@@ -5,7 +5,8 @@ export const StartButton = (props) => {
         <div className="d-flex flex-row justify-content-center container body RowStartButtonDims">
             <button
                 className="btn btn-success btn-lg"
-                onClick={props.onClick}>
+                onClick={props.onClick}
+                disabled={props.disabled}>
                 {props.value}
             </button>
         </div>
